renaming style files
