+++ Frontend/components/goalList.js
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import useGlobalHook from '../store';
 import apiRootUrl from '../utils/apiRoute';
 import GoalCard from './goalCard';
-import { goalList } from '../styles/goalList';
+import { goalList } from '../styles/styles-goalList';
 
 const renderGoals = goals => {
   if (!goals || goals.length === 0) return;
