+++ Frontend/styles/styles-goalCard.js
