+++ Frontend/styles/styles-newGoalModal.js
